requirements.txt
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.2
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4
